make frame counter 64bit
that's more useful since it can be used as macropulse number

+++ include/ChimeraTK/ControlSystemAdapter/MappedImage.h
@@ -85,7 +85,7 @@ namespace ChimeraTK {
     ImgFormat image_format{ImgFormat::Unset};
     ImgOptions options{ImgOptions::RowMajor};
     /// frame number/counter
-    uint32_t frame = 0;
+    uint64_t frame = 0;
   };
 
   class MappedImage;
